+++ src/Foundation/RelatedContentTagging/code/Models/RelatedTaggableContent.cs
@@ -4,6 +4,6 @@ namespace Hackathon.Boilerplate.Foundation.RelatedContentTagging.Models
 {
     public class RelatedTaggableContent : StringContent
     {
-        public double[] Vector { get; set; }
+        public float[] Vector { get; set; }
     }
 }
